replaced check with check2
