+++ cli/src/commands/index.ts
@@ -1,4 +1,5 @@
 export * from "./contract-diff-command";
-export * from "./check-command";
 export * from "./download-code-command";
-export * from "./check-command-2";
+export * from "./check-command";
+export * from "./storage-change-command";
+export * from "./storage-snapshot-command";
